+++ app/internal/config/buildkit_options.go
@@ -40,8 +40,16 @@ func (c BuildkitControlRequestBodyConfig) toPolicy(build BuildRequestBodyConfig)
 }
 
 func (c BuildkitSolveRequestBodyConfig) toPolicy(build BuildRequestBodyConfig) buildkitproxy.SolvePolicy {
+	var execDigests map[string]struct{}
+	if len(c.AllowedExecDigests) > 0 {
+		execDigests = make(map[string]struct{}, len(c.AllowedExecDigests))
+		for _, digest := range c.AllowedExecDigests {
+			execDigests[digest] = struct{}{}
+		}
+	}
 	return buildkitproxy.SolvePolicy{
 		Allow:                     c.Allow,
+		AllowedExecDigests:        execDigests,
 		AllowHostNetwork:          build.AllowHostNetwork,
 		AllowRemoteContext:        build.AllowRemoteContext,
 		AllowRunInstructions:      build.AllowRunInstructions,